requirements.txt
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
pillow>=9.0.0
pyscreenshot>=3.0.0
